+++ src/core/prelude/i18n/helpers.ts
@@ -10,8 +10,8 @@ import log from 'core/log';
 
 import langPacs, { Translation, PluralTranslation } from 'lang';
 
-import { locale, pluralizeMap } from 'core/prelude/i18n/const';
-import type { PluralizationCount } from 'core/prelude/i18n/interface';
+import { locale } from 'core/prelude/i18n/const';
+import type { I18nOpts, PluralizationCount } from 'core/prelude/i18n/interface';
 
 const
 	logger = log.namespace('i18n');
@@ -40,6 +40,8 @@ export function i18nFactory(
 		throw new ReferenceError('The locale for internationalization is not defined');
 	}
 
+	const pluralRules: CanUndef<Intl.PluralRules> = getPluralRules(resolvedLocale);
+
 	return function i18n(value: string | TemplateStringsArray, params?: I18nParams) {
 		if (Object.isArray(value) && value.length !== 1) {
 			throw new SyntaxError('Using i18n with template literals is allowed only without variables');
@@ -51,7 +53,7 @@ export function i18nFactory(
 			translateValue = langPacs[resolvedLocale]?.[correctKeyset ?? '']?.[key];
 
 		if (translateValue != null && translateValue !== '') {
-			return resolveTemplate(translateValue, params);
+			return resolveTemplate(translateValue, params, {pluralRules});
 		}
 
 		logger.error(
@@ -59,7 +61,7 @@ export function i18nFactory(
 			`Key: ${key}, KeysetNames: ${keysetNames.join(', ')}, LocaleName: ${resolvedLocale}, available locales: ${Object.keys(langPacs).join(', ')}`
 		);
 
-		return resolveTemplate(key, params);
+		return resolveTemplate(key, params, {pluralRules});
 	};
 }
 
@@ -85,9 +87,9 @@ export function i18nFactory(
  * console.log(examplePluralize); // '5 products'
  * ```
  */
-export function resolveTemplate(value: Translation, params?: I18nParams): string {
+export function resolveTemplate(value: Translation, params?: I18nParams, opts: I18nOpts = {}): string {
 	const
-		template = Object.isArray(value) ? pluralizeText(value, params?.count) : value;
+		template = Object.isPlainObject(value) ? pluralizeText(value, params?.count, opts.pluralRules) : value;
 
 	return template.replace(/{([^}]+)}/g, (_, key) => {
 		if (params?.[key] == null) {
@@ -117,15 +119,21 @@ export function resolveTemplate(value: Translation, params?: I18nParams): string
  * console.log(result); // '{count} products'
  * ```
  */
-export function pluralizeText(pluralTranslation: PluralTranslation, count: CanUndef<PluralizationCount>): string {
+export function pluralizeText(
+	pluralTranslation: PluralTranslation,
+	count: CanUndef<PluralizationCount>,
+	rules: CanUndef<Intl.PluralRules>
+): string {
 	let normalizedCount;
 
 	if (Object.isNumber(count)) {
 		normalizedCount = count;
 
 	} else if (Object.isString(count)) {
-		if (count in pluralizeMap) {
-			normalizedCount = pluralizeMap[count];
+		const translation = pluralTranslation[count];
+
+		if (translation != null) {
+			return translation;
 		}
 	}
 
@@ -134,18 +142,59 @@ export function pluralizeText(pluralTranslation: PluralTranslation, count: CanUn
 		normalizedCount = 1;
 	}
 
-	switch (normalizedCount) {
+	const
+		pluralFormName = getPluralFormName(normalizedCount, rules),
+		translation = pluralTranslation[pluralFormName];
+
+	if (translation == null) {
+		logger.error(`Plural form ${pluralFormName} doesn't exist.`, `String: ${pluralTranslation[0]}`);
+		return pluralTranslation.one;
+	}
+
+	return translation;
+}
+
+/**
+ * Returns the plural form name for a given number `n` based on the specified pluralization rules.
+ * Otherwise will be used default set of rules.
+ *
+ * If a `rules` object implementing `Intl.PluralRules` is provided, it will use that to determine the plural form.
+ * Otherwise, it will fall back to a custom rule set:
+ * - Returns 'zero' for `n === 0`.
+ * - Returns 'one' for `n === 1`.
+ * - Returns 'few' for `n > 1 && n < 5`.
+ * - Returns 'many' for all other values of `n`.
+ *
+ * @param n - The number to evaluate for pluralization.
+ * @param rules - Plural rules object. If undefined, a default rule set is used.
+ */
+export function getPluralFormName(n: number, rules?: CanUndef<Intl.PluralRules>): keyof Required<PluralTranslation> {
+	if (rules != null) {
+		return <keyof PluralTranslation>rules.select(n);
+	}
+
+	switch (n) {
 		case 0:
-			return pluralTranslation[3];
+			return 'zero';
 
 		case 1:
-			return pluralTranslation[0];
+			return 'one';
 
 		default:
-			if (normalizedCount > 1 && normalizedCount < 5) {
-				return pluralTranslation[1];
+			if (n > 1 && n < 5) {
+				return 'few';
 			}
 
-			return pluralTranslation[2];
+			return 'many';
+	}
+}
+
+/**
+ * Returns an instance of `Intl.PluralRules` for a given locale, if supported.
+ * @param locale - The locale for which to generate plural rules.
+ */
+export function getPluralRules(locale: Language): CanUndef<Intl.PluralRules> {
+	if ('PluralRules' in globalThis['Intl']) {
+		return new globalThis['Intl'].PluralRules(locale);
 	}
 }