+++ CHANGELOG.md
@@ -29,6 +29,15 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 * Conduct refactoring and optimization of the module `core/async`
 * Various performance optimizations
 
+## v4.0.0-alpha.46 (2024-09-25)
+
+#### :boom: breaking change
+
+* `core/prelude/i18n/helpers`
+  * changed `i18n` translations format.
+  * added `intl` support for pluralization.
+  * now `i18n` prefer to use `intl` api for pluralization if it's possible, otherwise fallback to old plural form logic.
+
 ## v4.0.0-alpha.45 (2024-09-09)
 
 #### :bug: Bug Fix